Added YQPkgSecondaryFilterView base class
For sharing the functionality for the views containing a secondary filter
